Fix more audio effects auto tests
commit ec00db4ed9374aaab44f976777d9e3f073effe47 broke the energy probe
used by auto tests to verify that an effect is actually enabled.
The probe is based on the Visualizer which now scales the capture to maximise
the signal level: this prevents from detecting level changes between subsequent captures.
Playing sound at a lower volume moves the signal level out of the range where
scalling occurs and makes comparison possible again.

Change-Id: Ib56dec6e619c6d778dcc3ca6453091bf6136f290