Fix deadlock in HTTP authentication within synchronous request.
When HTTP authentication results in the user being prompted for a username
and password, this is handled asynchronously in the UI thread. However, we
can't post the result back to the WebKit thread in the normal way because
WebKit does not pump its message queue during synchronous requests.

Instead, we call WebKit directly from the UI thread and let it handle the
threading. This is safe to do because we're using thread-safe STL and
Chrome objects, not WebCore objects.

Change-Id: I79c21510186c2352f842c4ff4311687ec40ba45f