Have download manager give more details when downloads are paused
This change introduces more detailed reporting of why downloads are paused in
the download manager.

First, it adds new, more detailed status constants in android.provider.Downloads
for when a download is paused, in place of the old catch-all
STATUS_RUNNING_PAUSED.  I've eliminated some dead code there as well to ease the
change.

Second, in the public API, it changes COLUMN_ERROR_CODE to COLUMN_REASON, which
now reports useful info for STATUS_PAUSED in addition to STATUS_FAILED.  It adds
some new PAUSED_* constants for reason values when a download is paused, and
adds code to generate these values appropriately.

Change-Id: Ie617d1c2c59357375e19466086923ace5e2fb122