Added noop unregisterReceiver to complement existing noop registerReceiver.
Without this addition, code that is being tested using an isolated context
can cause an exception if it unregisters a previously registered broadcast
receiver.  This is because the isolated context never actually registered
the receiver in the first place.  The fix is to make sure the isolated
context is consistent in ignoring both recevier registration and
unregistration calls.

Change-Id: Ie0ba6f4bb10f5248704327a0ffc8e37ee8b71ae2