Make View transformation matrix updating more explicit.
The original code depends on side-effects of hasIdentityMatrix()
to update the View transformation matrix.  This change adds an
explicit updateMatrix() method to make it more explicit when
we are relying on the matrix having been updated.  Among other
things, getInverseMatrix() did not update the matrix so the
caller had to know to call hasIdentityMatrix() first (which was
always the case, but still potentially error-prone).

Change-Id: I16d56f60a785a15c65f73e57b9aded9e4e3bca55