Fix Looper leaks in MediaRecorderStressTest.
The test was failing periodically due to too many files being open.
This change attempts to resolve the problem on the theory that
signaling pipe file descriptors are being leaked due to the large
number of Looper instances created during the test run.

However, it's still possible there are other leaks elsewhere.

Change-Id: I71f9f12d21605c47c9217c72c51e6c768142ce10