Handle native library deletion in scanPackage
The deletion of native libraries was initially added to
FileInstallArgs.cleanUpResourcesLI() as a way to get rid of old native
libraries during an upgrade, but it runs well after scanPackage unpacks
the new native libraries. scanPackage now removes old libraries before
unpacking the new ones, so we don't need this code anymore.

Bug: 3087739
Change-Id: I54aca830ec34d6440ba22f117d55aa3107bf5b75