Provide default error strings in BrowserFrame.onReceivedError().
This method is called from both the Android and Chrome HTTP stacks.
We want the same user-facing behaviour in each case, but the Chrome
code does not have easy access to the same error message strings.

Therefore, I'll have the Chrome HTTP stack provide empty strings
for the descriptions. (See change I9042a2ee in external/webkit).

This CL makes BrowserFrame check for empty or null strings and use
the standard error messages instead.

The strings were package-private, so I did a little cleaning up
to make them public (but still hidden from the API). Using a switch
statement instead of an array because it's more robust and should
be comparably efficient.

Bug: 3172265
Test: Load http://google.com:81 to force a timeout error, check
that the correct message is displayed.

Change-Id: I7dec8dffe8ad9e513fdb90877e0b466e707174a8