make tap initiated scrolls more reliable
Programmatic scrolls fail when the scroll has not had a chance
to start. In this case, the current position equals the last
position; the logic erroneously cancels the scrolls, assuming
that in this case the scroll has finished.

Also, new scrolls should cancel the old velocity whether they
are animated or not.

This has a companion change in external/webkit, which is not
required, but makes tap initiated scrolls work more reliably.

bug:1331125
bug:3099569
Change-Id: I80ae8e8cdefce2d8d09532f79d1b2e1ae4738fa6