Use apn types instead of host when sending notifications.
Previously we were passing the apn host not the type and this
caused us to sometimes not send out network change notifications.

Also rename of getActivceApn to getActiveApnHost to make
the method clearer on what is being returned and add some debug.

A future change will rename ApnSetting.apn to ApnSetting.host
and getActiveApnString.

bug: 2995554
Change-Id: I0c405b0f823f62c9596968c8516b820d9f8a94f2