Add reason field to DEACTIVATE_DATA_CALL request.
Commit 7838d79a719337af5ac9dd92e58795606f084903 introduced a side effect that
PPP and MIP cannot be disconnected gracefully in case of 1X data calls.
The new disconnect reason field in DEACTIVATE_DATA_CALL will help RIL and Modem
to know whether the disconnect was done for modem shutdown or not. And
the followings can be achieved at the same time.

- Graceful PPP termination
- MIP de-registration
- RF Release Order with "Power Down De-registration" if modem is powered off.

(Also cleaned up the deprecated functions for setup/deactivate default PDP)

Change-Id: If79842ad05827326b6994e1148a6ae68b4467c45